add about page footer
